openssh: support specification of SSHD_OPTS
Install /etc/default/ssh defining an empty SSHD_OPTS.
Modify the sshd unit to read this as an EnvironmentFile,
using the "-" modifier so that errors do not prevent
sshd from being started. Invoke sshd with SSHD_OPTS.
Modifying this simple defaults file is arguably easier
to maintain than changes to the extensive and protean
sshd_config.